add DeviceRegistrationServiceTest for registerDevice, add deviceName to DeviceRegisterResponseDto, add DeviceRegistrationVerificationData, add deviceTypeName to device name if name exists, add TestJwtBuilder

+++ src/main/java/com/iplion/mesync/cloud/config/RegistrationProperties.java
@@ -15,7 +15,7 @@ public record RegistrationProperties(
     @NotNull Duration inviteCooldown,
 
     @NotNull Duration registrationTtl,
-    @Min(1) int registrationAttempts
+    @Min(1) long registrationAttempts
 
 ) {
 }
